+++ src/components/Header/Header.tsx
@@ -8,7 +8,7 @@ import {
   StyledTitleDiv,
 } from '@components/Header/Header.styled';
 import * as Text from '@components/commons/Text/Text';
-import CartLengthButton from '@components/Header/CartLengthButton/CartLengthButtonButton';
+import CartLengthButton from '@components/Header/CartLengthButton/CartLengthButton';
 
 const Header = () => {
   return (
